+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ## [3.3.0](https://github.com/BLKSerene/Wordless/releases/tag/3.3.0) - ??/??/2023
 ### 🎉 New Features
+- Settings: Add Settings - Measures - Readability - Flesch Reading Ease
 - Utils: Add khmer-nltk's Khmer sentence tokenizer, word tokenizer, and part-of-speech tagger
 - Utils: Add PyThaiNLP's perceptron part-of-speech tagger (Blackboard)
 - Utils: Add spaCy's Korean sentence recognizer, word tokenizer, part-of-speech tagger, lemmatizer, and dependency parser
@@ -28,12 +29,14 @@
 
 ### ✨ Improvements
 - Utils: Update Wordless's sentence and sentence segment splitters
+- Work Area: Update Profiler - Readability - Flesch Reading Ease
 
 ### 📌 Bugfixes
 - Work Area: Fix Concordancer - Generation Settings - Width unit - Character
 
 ### ❌ Removals
 - Utils: Remove PyThaiNLP's perceptron part-of-speech tagger (LST20)
+- Work Area: Remove Profiler - Fernández Huerta's Readability Score / Szigriszt's Perspicuity Index
 
 ### ⏫ Dependency Changes
 - Dependencies: Add khmer-nltk